refactor: move supplier info into a dropdown card
