+++ packages/apps/stock/src/components/templates/supplierArrival/SupplierArrivalHeader/SupplierArrivalHeader.tsx
@@ -16,58 +16,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, {useCallback, useMemo, useState} from 'react';
+import React from 'react';
 import {StyleSheet, View} from 'react-native';
-import {Button, LabelText, useThemeColor} from '@axelor/aos-mobile-ui';
-import {
-  formatDate,
-  useDispatch,
-  usePermitted,
-  useTranslator,
-  useTypes,
-} from '@axelor/aos-mobile-core';
+import {LabelText} from '@axelor/aos-mobile-ui';
 import {StockMove} from '../../../../types';
 import {StockMoveHeader} from '../../../organisms';
-import {updateSupplierArrivalShipmentDetails} from '../../../../features/supplierArrivalSlice';
-import {SupplierArrivalShipmentDetailsPopup} from '../../supplierArrival';
 
 const SupplierArrivalHeader = ({supplierArrival}: {supplierArrival: any}) => {
-  const I18n = useTranslator();
-  const Colors = useThemeColor();
-  const dispatch = useDispatch();
-  const {StockMove: StockMoveTypes} = useTypes();
-  const {readonly} = usePermitted({
-    modelName: 'com.axelor.apps.stock.db.StockMove',
-  });
-
-  const [isEditPopupVisible, setEditPopupVisible] = useState(false);
-
-  const showEditButton = useMemo(
-    () =>
-      supplierArrival.statusSelect !== StockMoveTypes?.statusSelect.Realized &&
-      !readonly,
-    [
-      StockMoveTypes?.statusSelect.Realized,
-      readonly,
-      supplierArrival.statusSelect,
-    ],
-  );
-
-  const handleEditConfirm = useCallback(
-    (ref: string | undefined, date: string | undefined) => {
-      setEditPopupVisible(false);
-      dispatch(
-        updateSupplierArrivalShipmentDetails({
-          id: supplierArrival.id,
-          version: supplierArrival.version,
-          supplierShipmentRef: ref ?? null,
-          supplierShipmentDate: date ?? null,
-        }),
-      );
-    },
-    [dispatch, supplierArrival.id, supplierArrival.version],
-  );
-
   return (
     <View>
       <StockMoveHeader
@@ -90,39 +45,7 @@ const SupplierArrivalHeader = ({supplierArrival}: {supplierArrival: any}) => {
         {supplierArrival.origin == null ? null : (
           <LabelText iconName="tag-fill" title={supplierArrival.origin} />
         )}
-        <LabelText
-          title={I18n.t('Stock_SupplierShipmentRef')}
-          value={supplierArrival.supplierShipmentRef ?? '-'}
-        />
-        <LabelText
-          title={I18n.t('Stock_SupplierShipmentDate')}
-          value={
-            supplierArrival.supplierShipmentDate
-              ? formatDate(
-                  supplierArrival.supplierShipmentDate,
-                  I18n.t('Base_DateFormat'),
-                )
-              : '-'
-          }
-        />
-        {showEditButton && (
-          <Button
-            title={I18n.t('Stock_FillShipmentDetails')}
-            width="100%"
-            iconName="pencil-fill"
-            iconSize={16}
-            color={Colors.infoColor}
-            onPress={() => setEditPopupVisible(true)}
-          />
-        )}
       </View>
-      <SupplierArrivalShipmentDetailsPopup
-        visible={isEditPopupVisible}
-        defaultRef={supplierArrival.supplierShipmentRef}
-        defaultDate={supplierArrival.supplierShipmentDate}
-        onConfirm={handleEditConfirm}
-        onCancel={() => setEditPopupVisible(false)}
-      />
     </View>
   );
 };
