+++ packages/apps/stock/src/components/templates/supplierArrival/index.tsx
@@ -28,5 +28,6 @@ export {default as SupplierArrivalMovementIndicationCard} from './SupplierArriva
 export {default as SupplierArrivalOriginInput} from './SupplierArrivalOriginInput/SupplierArrivalOriginInput';
 export {default as SupplierArrivalSearchLineContainer} from './SupplierArrivalSearchLineContainer/SupplierArrivalSearchLineContainer';
 export {default as SupplierArrivalShipmentDetailsPopup} from './SupplierArrivalShipmentDetailsPopup/SupplierArrivalShipmentDetailsPopup';
+export {default as SupplierArrivalShipmentDropdownCard} from './SupplierArrivalShipmentDropdownCard/SupplierArrivalShipmentDropdownCard';
 export {default as SupplierArrivalTrackingNumberSelect} from './SupplierArrivalTrackingNumberSelect/SupplierArrivalTrackingNumberSelect';
 export {default as SupplierProductInfo} from './SupplierProductInfo/SupplierProductInfo';
